Update patch set 4
Patch Set 4:

Maybe we're looking at it a different way

 start_demand = 5
 ready = 1
 min_ready = 3

means to me that we want to request 4 from the allocator (that covers our demand, the minimum we want ready and what we've got)

if say

 start_demand = 1
 ready = 0
 min_ready = 10
 capacity = 2

then we don't want to request 11 from the allocator, as we know there's only 2 slots available.  we should just request the 1 node we really need.

maybe we need to tighten the definition of what min-ready is?

Patch-set: 4