Update patch set 3
Patch Set 3: Code-Review-1

(3 comments)

this is not like a lot of other mock i've seen.  is there some reason why simple use of the decorator will not work here?  I think it will remove a lot of the code, and we shouldn't need to modify the imports of the cmd file

Patch-set: 3
Label: Code-Review=-1