Update patch set 2
Patch Set 2: Code-Review-1

A comment could be useful to explain that a new node is still created in the failed deletion case due to the failure being to communicate with Jenkins (and not that nodepool failed). That is, the node was still deleted and a new one created.

The flip side could be true too. What happens when nodepool fails to delete a node for some reason (provider disappears etc). Do we expect a new node or the old one?

Patch-set: 2
Label: Code-Review=-1