Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

We need to document the new ipv6-preferred option.

I also would like to see tracking of both IP's, but I think we probably need a new table to track all of the IP's assigned to a node so it's probably another change.

Patch-set: 2
Label: Code-Review=-1