Run tasks in worker threads
The task manager was designed under the assumption that API requests
are relatively quick and should be serialized to avoid hitting a
rate limit.  We have found that sometimes that is the case, however,
other times, API requests may take, for example, 30 seconds to complete.
Even with a very conservative API rate target of 1 request per second,
this means nodepool is 30 times slower than necessary.

This change alters the task manager so that it causes an API request
to _start_ at the specified rate limit, but does not account for when
it completes.  It does this by using the task manager thread to
serialize the start of requests, but the requests themselves are now
performed in the calling thread.

Change-Id: I025250494e7d88b82fea4514f74f204138a14d94