Update patch set 2
Patch Set 2: Code-Review-1

I'm not a fan of the group/groups being the top of the metavalues. If I were using nodepool and say those in the cloud I'd have no idea what they mean. However this has been dictated by ansible, not this change, so it is (somewhat) necessary.

I wonder, however, if we should document or toggle the feature of adding ansible information so that if somebody comes looking for explanations of the metavalues they'll find a reference? (if you disagree, then +2 on code)

Patch-set: 2
Label: Code-Review=-1