Deal with failures that succeed
Some clouds regularly return 500 level errors, but then go on to
process the request and the server requested shows up happily. Since we
already have a timeout loop on server creation, let that loop handle
waiting for the server object to show up in the cloud.

This will need a follow on patch to deal with marking servers in this
state that never come back for deletion from the database... they'll be
crufty rows ultimately otherwise.

Change-Id: Idb6ca5ac5e0eb1ce79b27367920331faffce6396