Update patch set 1
Patch Set 1: Code-Review-1

(2 comments)

The user experience here is a bit odd. If you do an alien-image-list and it fails, you just get an empty table. You might assume this means there are no alien images. We should raise or print the error to the user in some form.

Patch-set: 1