Update patch set 3
Patch Set 3:

The problem is that we are running os.path.join() over and over on the same string. Basically doing os.path.join('/foo/bar', '/foo/bar/baz') constantly. Before we write to the db, after we read from the db and so on. What will happen in the case of relative paths is os.path.join('foo/bar', os.path.join('foo/bar', 'foo/bar/baz') and you end up with the wrong path. The cases where this can happen are few, but it makes the code incredibl hard to follow because this sort of nested path joining is what we are doing.

Patch-set: 3
Reviewer: Gerrit User 4146 <4146@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0