Use shade to get the client objects
As an initial step in the transition to shade, just replace the getting
of the client objects with shade. This adds support for but does not
require using a clouds.yaml file and named clouds. It also gives us
access to other client libraries, such as neutron, without structural
changes, although the attempt here is to not change anything logically
or move the ownership of any code tasks. The one exception is the
removal of using the novaclient extension for finding the neutron
networks, since shade neither has a facility to passing novaclient
extensions in as parameters, nor does it want to grow one.

Change-Id: Iba84b8d578efa8cb7f5af85a5ffbc97f945a47c9