WIP: Use shade to get the client objects
As an initial step in the transition to shade, just replace the getting
of the client objects with shade. This adds support for but does not
require using a clouds.yaml file and named clouds. This is an opt-in
feature and is not required by people using nodepool. But, having a
clouds.yaml on a server allows the same cloud configuration to also be
used by other tools, such as ansible and python-openstackclient, so
while not required, it's a nice to have.

Using shade to get the client connection handles also gives us
access to other client libraries, such as neutron, without structural
changes, although the attempt here is to not change anything logically
or move the ownership of any code tasks. The one exception is the
removal of using the novaclient extension for finding the neutron
networks, since shade neither has a facility to passing novaclient
extensions in as parameters, nor does it want to grow one.

Change-Id: Iba84b8d578efa8cb7f5af85a5ffbc97f945a47c9