Use shade for image uploads
The most important thing for dib-nodepool is for nodepool to be
able to upload images to both clouds. Sadly this isn't particularly
straightforward, so we let shade handle the heavy lifting.

Note: image uploading remains a blocking call in v1, but non-blocking in
v2 - so there is some opportunity for unhappiness here.

Change-Id: I96b7529510ca24d3b164b2d49da1d0a02b808fa5