Use shade API for flavors
This isn't perfect, in that there is an unnecessary thread task created
for something that is in almost all cases going to be a cache hit.
However, it's a good line in the sand to see if we can figure out the
proper API between nodepool and shade such that even the task on the
nodepool side can go away, but we can still stay within API rate limits,
without making the shade API crazypants for other things.

Depends-On: I65a7b1bf6818f6b40b42bd2fa4b560f127c5589b
Change-Id: I62709ec8586eb4299aa82187bfd0260ba35cb4ea