Use PriorityQueue to guarantee key tasks are coming first.
Under load, we see that the key polling tasks to check
for server list, ips, etc... are not running on the
desired interval, because they are queued after operational
tasks such as create servers and ips. Use a PriorityQueue
to guarantee that these tasks will run when desired.
As these tasks do also benefit from caching, we are sure
that these won't be spamming the task queue, as they
will only be running after the cache has expired.

Change-Id: I19e12cd7c8d2be046ad500ac408bf6ec30e83209