Update patch set 8
Patch Set 8: Code-Review-1

(3 comments)

I have inline suggestions about how to change the data structure to, I think, better represent what we're modeling while reducing repitition and potential for operator error.

I think we should make the image type (and related options) attributes of providers, and when building diskimages, we should take the union of all type/attribute combinations and build all of them for every image.

That might cause us to build an image for a provider even if that image is not used with that provider, however, in general, we want to use the same images everywhere.  That's not the case with the tripleo cloud (which has unique images), but I don't think the extra computation is going to kill us.

If we decide that we want to avoid the above de-optimization, we could, instead of generating the union of all types for all images, simply calculate which images are used on which provider, and generate the appropriate images for that provider.  That means that if we create a qcow image for provider B, but later add that image to provider A, nodepool will need to know to create a new image or convert it.  If that's easy to do, that's actually probably the better solution (as compared to the union idea above).

If shade can provide the appropriate information (by either the a-priori information in os-client-config or reading it locally from clouds.yaml) that would actually probably be a better source of information than adding it to nodepool.  I'm not certain if we want to start by doing that, or just add it to nodepool now and move to shade later.

Patch-set: 8
Label: Code-Review=-1