Update patch set 3
Patch Set 3:

I cant think of a possible use case where not 0.10 is desired, but there are ones where 0.20 does not work. I was also considering just proposing this as the default for dib for this reason...

IMO I do not think hard-coding this should be an issue unless we can come up with a use case where 0.10 does not work.. and if we can then we should be making this configurable per image.

Patch-set: 3
Label: Verified=0