Update patch set 8
Patch Set 8:

The coverage report for this change is 1% worse than for the most recently merged change to the nodepool repo:

http://logs.openstack.org/89/170289/1/check/nodepool-coverage/8cff29a/cover/

That conflicts with the goal you state in the commit message. If the plan is to build on this to increase test coverage, then I'm +2, but as it stands this change alone seems to slightly regress test coverage instead.

Patch-set: 8
Label: Verified=0