Update patch set 2
Patch Set 2: Code-Review-1

I'm not convinced that this is the best approach -- for example when you .stop() a ProviderManager that is not going to call stop() for the embedded SlowProviderManager

in general nesting these seems like breaking the abstraction they are providing.

one alternative option might be a priority queue.   Or have TaskManager provide multiple queues, and you submitTask() to a "slow" or a "fast" queue.  both these seem to fit the existing abstractions better

Patch-set: 2
Label: Code-Review=-1