Update patch set 4
Patch Set 4:

> Not sure how I feel about this patch, we haven't really included distro-specific dependencies and OpenStack Infra itself uses Ubuntu, I worry that specifically including RH/CentOS implies otherwise. I'd like to see some thoughts from other cores.

I think we should list both.

My team are not using Ubuntu so I don't know what apt-get packages that are needed to get nodepool going on a fresh Ubuntu system but I think it would be useful to list Distro instructions for distros we know work.

FWIW we the Linux Foundation are deploying nodepool on Centos 7.1 images for some of our projects.

My hope is to make the install process as easy as it can be for future new deployments, I walked through the process of getting nodepool running myself for the first time 2 weeks ago and felt the documentation was confusing in a few areas which is why I'm now opening patches for parts that I got stuck on. Hoping to clarify it for the next guy.

Patch-set: 4
Label: Verified=0