Update patch set 8
Patch Set 8: Code-Review-1

I think we should do this without changing either the TaskManager interface (this patch does not change it, so that's good), or the Task interface (this patch does change that).

That's because when we switch to shade, it will be shade that is providing the tasks, so we can't count on them indicating internally that they should go to one queue or the other.

[This is a system that we should never need, except that we do need it because one of our providers has a production error that they seem reluctant to fix.  This is not needed in any reasonably modern and up to date openstack deployment.  It's certainly not part of the OpenStack api or any expected behavior.  Therefore it doesn't belong in shade, it belongs entirely in nodepool; and it's reasonable for nodepool to make decisions about how and when certain tasks should be executed (we are making that part of the shade API).]

So anyway, I think that means that instead of adding the information to tasks, we actually just need to put something in the nodepool TaskManager that does something like:

  if isinstance(task, CreateServerTask):
    ...

Or maybe generalize that into having a list of "slow tasks" and checking whether the task is in that list.  When we add shade, we can add shade's version of those tasks into the list.

Patch-set: 8
Label: Code-Review=-1