Document running Nodepool for testing
Write a small document explaining two methods for testing Nodepool.
First running the functional tests with tox and setting up the database
for that and second running Nodepool against a Devstack deployed cloud.

This should help people wanting to quickly use Nodepool in a non
production environment without needing to read all of the configuration
documentation.

Change-Id: Icbf19a25a8b555f1289ff9efb368e9e5cd4740ab