Consume os-client-config directly
In order for us to notice config changes in clouds.yaml, rather than
consuming it indirectly via shade, create the cloud config directly
and pass it in to the shade factory function.

There would not be a huge win in teaching shade to notice and update
itself at the moment, as the update processing happens in a different
place and there isn't really a great way to have the check ask the old
provider manager to see if its cloud config has changed.

Change-Id: Ie4cf44adefda637b5987dd9d39ec398786086720
Depends-On: Icdd9acede81bc5fba60d877194048e24a62c9e5d