Update patch set 2
Patch Set 2:

I have previously said that tests should be added as well, but I am failing to see what kind of tests are appropriate for this change. It is not changing behavior. And none of the other timeout settings are being tested, afaik (not a good excuse, I know). So I am looking for pointers. Are we supposed to add full-blown scenario test to check if the setting this timeout would make nodepool kill a node at a correct time? Or just checking if the variable is getting set given a config? Is it an overkill to add such tests?

Patch-set: 2
Reviewer: Gerrit User 7979 <7979@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0