Update patch set 22
Patch Set 22:

No, it's different issue thatn the workers spec. This is a workaround for when cloud api times are slow. We had this issue in the past, where hp cloud api times were over 20 seconds per call, and redistributing the tasks in two queues was really useful. At the moment api calls have reduced times for 2-3 seconds, so this change is less prioritary, but i'm keeping that up to date with the hope it lands, because it provides improvement anyway.

Patch-set: 22
Reviewer: Gerrit User 6133 <6133@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0