Update patch set 2
Patch Set 2:

I agree that we should add a test. Even though we are not changing any functionality, we are adding a configuration option to a previously-untested code path, and I would like to see a test added that ensures that the timeout triggers an appropriate delete while we are modifying the code path around it.

Patch-set: 2
Reviewer: Gerrit User 6039 <6039@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0