Update patch set 1
Patch Set 1:

In my opinion understanding the deep context of what is happening is a debug mode. The info should be informative without being overly verbose.

An error is already raised when the command itself fails. I think that covers the high level information to know something has failed and if they need to dig the operator would turn on debugging.

I'd like to see it changed to debug, but also happy to be out voted :-)

Patch-set: 1