Allow to delete nodes with missing target or provider
When a target or a provider is incorrectly removed
from nodepool, it can leave nodes that can be moved
to delete state, but then they cannot be successfully
deleted because missing provider or target.

Make this code more resilient, to allow the delete of
these nodes despite of this. If not, the only alternative
is to manually delete the nodes from the database.

It includes sanity checks to anywhere when provider or
target references are used (such as allocation calculations),
to ensure that everything is done against valid providers
and targets.

Change-Id: If4ec58e8beccb3f61b6dfa64de4b11910b2ba515