Update patch set 1
Patch Set 1:

I understand the argument, but here's problems in deploying nodepool daemon without the puppet files: Logging was one of the points that needed a default setting that is sane.

I examined other openstack projects and saw that most configuration tools are either integrated in the project or tools directory and saved it there.

If the nodepool is to move outside the openstack-infra, the application should come with all the necessary tools. If this solution is not acceptable, then how about moving the script into a tools directory in the root of nodepool?

Patch-set: 1
Reviewer: Gerrit User 8984 <8984@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0