Update patch set 7
Patch Set 7: Code-Review-1

This looks like its making the secrets file required? This could be painful for downstream users - they are forced to change their config layout immediately when we release this. I think it would be a lot nicer to them for us to add this as an optional feature.

Patch-set: 7