Update patch set 1
Patch Set 1:

There may be various options to make the setup even better, but I still think that my patch is a useful first step, as it prevents trying to exec dib with a broken enviroment, leading to the difficult to interpret error message shown in the commit message.

If, by not setting the directories properly, some needed scripts or elements can not be found, there should be a message about that during the execution of dib that should be much more useful to the user.

Patch-set: 1
Label: Verified=0