Fixed crash in exception handler
Due to a type mismatch problem, the exception handler it was itself
throwing an exception.

When the exception handler was attempting to write to stderr, it adds a
newline character to the exception message via

   e.message + '\n'

e is of type Exception.  The message field of objects of type Exception
**should** have a type str, however for some of the actual exceptions
thrown by this block of code this was not the case.  This causes the +
operator to fail at runtime.

For example, for an exception that had type
requests.exceptions.ConnectionError, the type of the message field was
requests.packages.urllib3.exceptions.ProtocolError, not str.

In addition to being incorrect, writing to stderr directly in an
exception handler is stylistically different from the rest of the
project.

To rectify this, I have removed the write to stderr and I have raised
the log level to 'exception'.

Change-Id: I9e980acb48726ae3a8242b579091e02fdf7bfd96