Update patch set 1
Patch Set 1: Code-Review-1

I kind of agree with ianw here. I am not sure why were writing to stderr but then doing a log.debug on this... maybe the correct fix is to remove the stderr write and up the log level to warning or similar (I dont think we should be getting exceptions in these methods in normal use).

Patch-set: 1
Label: Code-Review=-1