Update patch set 3
Patch Set 3:

Looking at the context of the code, I am actually thinking there is a 
good reason why not to raise an exception here.  In these functions
nodepool contacts all of its providers to get generate a list of aliens
or a list of alien images.  The try-except block is wrapped around the 
code that contacts a single provider.  If we run into difficutly 
contacting one provider I think the best thing to do is to inform
the user, skip the provider, and continue with the remainder.

This also got me thinking about the content of the log message.  In its
current state it is not informative at all.  In my most recent patch
set I have updated this message so that it indicates the name of the
provider(s) that cannot be properly contacted for some reason.

Patch-set: 3
Label: Verified=0