Update patch set 4
Patch Set 4:

Khai wrote:
> Check change Ia6ce75958bb1 : Darragh says "The only problem is that on python 3, these need to be byte strings, where as on python 2 they can be either"

Amended to pass b'' instead of plain ''.

> Also was wondering why not just use the python-jenkins library to do all this work?

My rational is that it is a very quick fix to have Nodepool supports recent version of Jenkins, or at least on my setup.   Upgrading the python-jenkins dependency is definitely wanted, but I have the feeling it is much more work.

Patch-set: 4