Update patch set 2
Patch Set 2: Code-Review-1

I hate to be negative-nancy here, but i think the intent of name-filter is not to select the "m1.large" type -- that is automatically deduced from the min-ram.

The thing is when rackspace offers something like two versions of m1.large with the same ram, but one pvhm and the other not (their "performance" nodes).  so this is to distinguish that

i.e.; having it as "performance", as in the documentation below, is probably better

Patch-set: 2
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1