Update patch set 14
Patch Set 14:

Well, that has been discussed several times from now, in the review and in the channel. It came from an agreement between jeblair, mordred, asselin... and myself. However, if you have concerns, i think we could talk about it on the channel at some point , or at the infra meeting.
Having that secure.conf is nothing different that we have on zuul for example, where we have the zuul.conf and the layout.yaml
Behaviour in nodepool will be pretty much the same. secure.conf will contain the credentials, and nodepool.yaml will define the releationship with cloud, jenkins... it will be no longer a configuration file.

Patch-set: 14
Label: Verified=0