Update patch set 14
Patch Set 14: Code-Review+2

Yep, we talked about this a few times, including during the recent sprint.  It is essential that we move the secret information out of the main config file.  That way it can be treated as "content" (it describes how to build images, which clouds to use for which images, etc.).  The nuts and bolts information about how to talk to clouds, jenkins., etc, should be in a different file.  Because Monty wants to re-use the same file for nodepool and the CLI, the cloud creds are in clouds.yaml.  We still need a place to put the remaining secret credentials, at least until we stop using jenkins.  Aside from clouds.yaml, most of the programs we use use the ini file format for credentials.  This matches that pattern.

Patch-set: 14
Label: Code-Review=+2