Stop all threads on SIGUSR1
I noticed SIGUSR1 is supposed to gracefully stop nodepool. Unfortunately
a lot of threads are never stopped and thus never stop.

[nodepoold.py]
Ignore SIGUSR1 while it is being dealt with. I had some side effect when
sending it twice in a row.

Teach the main loop to return whenever the pool is stopped after a
signal else nodepool never leaves when I we run it in non daemon mode
with 'nodepoold -d'.

[nodepool.py]

DiskImageBuilder can now be stopped. It seems it should inherit
TaskManager but I felt it is a different topic.

When Nodepool stop, also stop:
* Jenkins managers
* Cloud providers
* Gearman client

Add a bunch of debug log to notice what is going on.

Change-Id: Ic20f919df3fab834fb1a29df5c77a4af79e740a6