Update patch set 1
Patch Set 1: Code-Review-1

I agree on the test-case; but also...

SIGUSR1 handling seems to have been added in the initial commit (I7af197743cdf9523318605b6e85d2cc747a356c7) and I think the whole thing looks kind of odd?

I'm just guessing at the history, but the main loop sits waiting for a KeyboardInterrupt exception but never actually sets a SIGINT handler.  I wonder if the original SIGUSR1 catch was put in because the more logical (IMO) sigint handling was noticed to be not working when not run from a command-line because SIGINT is default ignored in this case, and there's no explicit handler setup?

So this behaviour isn't documented and quite possibly wrong, so I think I'd prefer to see this just catch SIGINT and use that as the exit code... but either way we should document it

Patch-set: 1
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1