Update patch set 1
Patch Set 1:

The plan is to test nodepool with a real cloud. In the openstack world that means using devstack. We could do devstack + otherthing to deploy nodepool but why not add a devstack plugin in order to deploy nodepool against the cloud?

Once we have this working we can then gate nodepool changes against their ability to build and image and boot a slave with that image.

Patch-set: 1
Label: Verified=0