Update patch set 1
Abandoned

This was not the change we needed. Because shade and os client config normalize the config this is a noop change for talking to keystoneclient. Instead we needed a fix to os client config to treat v2password auth type as regular generic password auth type.

Patch-set: 1
Status: abandoned