Update patch set 14
Patch Set 14: Code-Review-1

Don't hate me ... but maybe instead of the worker having a nodedb connection, the nodedb work should be done by nodepool and the appropriate things should be passed as parameters/return values to and from the workers over the gearman. That would make the system a little more decoupled and would make it easier to use network security between nodepool and its database.

Patch-set: 14