Update patch set 14
Patch Set 14:

I agree with what you're saying re: the workers should not end up with a nodedb connection, this is just my attempt at finding a path where we can end up there and not have a single mega patch. As per the spec, the plan was to move the image building bits into their own config file which the workers can read (rather than pass that data over the gearman bus). My plan was to define that config in the patch after the image deleting one in this series and then remove all the nodedb connection bits.

Patch-set: 14
Label: Verified=0