Update patch set 1
Patch Set 1:

It was recommended in IRC that we use openstack_cloud() from shade instead of calling the constructor directly. However it looks like openstack_cloud() is looking for an OpenStackConfig object (I'm guessing, judging by what methods it calls on it later), but the ProviderManager.provider.cloud_config in this context is a CloudConfig object. So it doesn't look to me like this will work, but I'm happy to be corrected by someone more experienced with this.

Patch-set: 1