tools: shell script to gracefully stop and wait
systemd default to stopping the service with SIGTERM which Nodepool
handles by abruptly exiting.  Instead one would want to send SIGUSR1 and
wait for the service to complete gracefully.

Add tools/nodepool-graceful-stop which can be used in systemd with:

    ExecStop=/usr/bin/nodepool-graceful-stop $MAINPID

Note: the script send SIGUSR1 for backward compatibility in case someone
copy paste the script and attempt to run it on a version not handling
SIGINT (which superseeds SIGUSR1 nowadays).

Signed-off-by: Moritz Muehlenhoff <mmuhlenhoff@wikimedia.org>
Change-Id: If07303d04aa83d15ee440c2c32541bb2fa66477d