nodepool: resolve package dependence issue
Add MySQL-python into requirements.txt to resolve the package
dependence issue for nodepool: When mysqldb is chosen as the
DB connector, sqlalchemy will try to import the module of MySQLdb.
So we need install the package of MySQL-python. Otherwise, DB
access may fail due to "ImportError: No module named MySQLdb".


Change-Id: Ie41b3610adf05a5ee1eda21b277642f01e265d3d