Update patch set 1
Patch Set 1:

Of course I neglected to point out that nodepool runs as the stack user when used as a devstack plugin. But that is why this is likely not something we should be doing here and instead update the job if required to run as the stack user.

Patch-set: 1