Update patch set 1
Patch Set 1: Code-Review-1

So as the code reads ATM, the !include only picks up a sibling file in the same directory (since it's using the directory from the loader object, which I guess is a file-like object of the main .yaml file?)

So that's probably fine, but I think just needs to be documented as being a relative path to the including file.

Queries: I guess voluptuous doesn't give us line data on the config files anyway, so that's not broken -- although it might be a little harder to pin down what it's talking about?

Does it nest?

Patch-set: 1
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1