Document setup uses root or well known cloud users
I though the 'setup' step was using the jenkins username I defined.
Turns out SnapshotImageUpdater.bootstrapServer() attempts 'root' then
well known cloud users (e.g: ubuntu, cloud-user...) which is supposed to
have sudo access.

The good point, is that there is no need for a sudo rule for the jenkins
user!

Mention in the scripts documentation that the setup script is connecting
as root/cloud users.

Change-Id: Ie3247d2c55009fcde8fd327f6c061ceb1e76d0ba