Sync to global requirements
Do an initial sync to global requirements, in preparation for
consuming them.

This is important because we are consuming other OpenStack projects
using global-requirements, such as diskimage-builder, and if we get
out of sync we run the risk of asynchronous breakage
(e.g. I0efa395bc66c9e50bfe4a13e971232ad9184ad48)

Change-Id: Ie493c69d9e1b50b1101a50962a8e03806ed1f120
Depends-On: Id500f5d922fd270c2f54bfd5e931bf62807eee12