Update patch set 1
Patch Set 1:

Now that I read more and look closely at the job logs I see the intent is to record the last image listing and last node listing that was made. This is fine, but the code fails due to permissions mismatches. We should correct that.

Also, the reason this had a 15 minute timeout was the longest total build time based on testing appeared to be ~13 minutes. This gave us a little bit over extra grace time but prevents the jobs from running all the way to some crazy devstack timeout. In this case it is 2 hours which maybe we should decrease to closer to an hour if we use this approach.

Patch-set: 1