Builders no longer use nodedb for diskimages
Currently we use the nodepool config object and nodedb database
to determine information about the diskimages we build. Start parsing
the config and stop using nodedb in builders.

This changes images to use a uuid for ID's since the primary key
ID which was previously used is no longer readable by the builders.

Conflicts:
	nodepool/nodepool.py

Co-Authored-By: Joshua Hesketh <josh@nitrotech.org>

Change-Id: Id1bb431839cde2bd7b42cf1f57460febd56849b3