Use shade for all OpenStack interactions
We wrote shade as an extraction of the logic we had in nodepool, and
have since expanded it to support more clouds. It's time to start
using it in nodepool, since that will allow us to add more clouds
and also to handle a wider variety of them.

Making a patch series was too tricky because of the way fakes and
threading work, so this is everything in one stab.

Change-Id: I423716d619aafb2eca5c1748bc65b38603a97b6a
Depends-On: Ic3345e7fb7e428227a9465bd07077e67d4f93244