Use shade for all OpenStack interactions (jeblair's version)
We wrote shade as an extraction of the logic we had in nodepool, and
have since expanded it to support more clouds. It's time to start
using it in nodepool, since that will allow us to add more clouds
and also to handle a wider variety of them.

Making a patch series was too tricky because of the way fakes and
threading work, so this is everything in one stab.

Change-Id: I69b5a0246d5d3c1de08334f0b80d1f3d36526bba
Depends-On: Ia8af343b17e05191e47dc7f3969aeaf19a443da6