Restore ability to run nodepoold with fakes
With OSC and shade patches, we lost the ability to run nodepoold
in the foreground with fakes.  This restores that ability.

The shade integration unit tests are updated to use the string
'real' rather than 'fake' in config files, as they are trying to
avoid actually using the nodepool fakes, and the use of the string
'fake' is what triggers there use in many cases.

Change-Id: Ia5d3c3d5462bc03edafcc1567d1bab299ea5d40f