Retry creating of floating IPs on failures
Currently nodepools deletes and restarts a new instance if the creation of a new
floating IP fails.  However, it might be that this is a temporary shortage of
available IPs, and in that case a lot of instances might be restarted
continuous.

This patch changes this behavior, and tries to create a new floating IP
repeatedly until it either succeeds or hits the default timeout of 600 seconds.

Change-Id: I9ba5ec717497daba64e6120da0bff2360078ca87