Cleanup floating ips even if server unavailable
cleanupServer() should handle deletion of floating IPs, keypairs and finally the
server itself. However, currently it works only if the server itself is still
running and thus included in the servers list.
If the server is no longer listed, the floating IP and keypair are never deleted,
because an exception is raised before these tasks.

This patch tries to delete a floating IP even if the server itself has not been
found. Also, if the server has been found and an error occurs when deleting the
floating IP or keypair, the server will still be deleted finally.

Change-Id: Ic11be567e5e8857abf4a37f4b13f9e4f33495b5c