Cleanup floating ips and keypairs even if server unavailable
cleanupServer() should handle deletion of floating IPs, keypairs and finally the
server itself. However, currently it works only if the server itself is still
running and thus included in the servers list.
If the server is no longer listed, the floating IP and keypair are never deleted,
because an exception is raised before these tasks.

This patch changes the order of tasks inside the method to ensure that floating
IPs and keypairs are deleted even if the server itself is not longer running.

Change-Id: Ic11be567e5e8857abf4a37f4b13f9e4f33495b5c