Update patch set 2
Patch Set 2: Code-Review+1

++ for things that make us perform less API requests.

Having to do some manual cleanup if we move out of a region doesn't seem like too big of an issue. The biggest concerns I have with never scaling down is the wasted resources (when load is low) or if we want to scale down the number of nodes we want in a cloud we will permanently be wasting FIPs. I could envision a set up where nodepool is smarter and maybe only performs FIP deletion if there is no demand for nodes, but IMO it isnt worth the complexity until we hit an issue with it.

Patch-set: 2
Reviewer: Gerrit User 10035 <10035@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1