node deletion delay is now configurable
The nodepool daemon is hardcoded to wait one minute before deleting a
node.  Apparently to let a Jenkins publisher to finish copying the
console output.  When such a system is not in use, that uneccessary
delay the used nodes and slowdown the pool refilling.

Introduce a new section 'nodepoold' in nodepool.yaml
Add the optional setting 'delete-delay'

The delay still fall back to the hardcoded value.

Also fixed a tiny unrelated doc issue.

Change-Id: I87afce1a1f3457e107b2802a67c49d5dbd164e4e