Write the cloud name into the provider info
So that we can do per-region mirrors with the name of the cloud
accounted for, write out a NODEPOOL_CLOUD variable. Also, to ensure this
doens't go south for people not using clouds.yaml, check to make sure
there is always a name parameter to get_one_cloud() and if there is not
one specified, use the provider name.

Change-Id: Ic6b72b5ecc52dad243d8be91469ab2b6073a4dc5