Add provider node stats
We dropped the ability to get # of <state> nodes per provider.
This was computationally expensive before; add it back in a simpler
form (not broken up by image type).  This will add another 30
metrics or so, but is worth it, and is still much less than the
hundreds we had before.

Change-Id: I371e2bc7e746c6eb00bd735edc009ef7b49af466