Update patch set 13
Patch Set 13: Code-Review-1

(3 comments)

Some comments inline. To get better coverage of the out of process builder command maybe we run a separate builder process in the devstack test using the devstack plugin? Then we don't cover the in process builder too well there but the unittests seem to do that for us so we should be good.

Patch-set: 13
Label: Code-Review=-1