Update patch set 1
Patch Set 1: Code-Review-1

(2 comments)

A couple of minor things inline. Greg mentioned adding tests so that if rebasing the image builder stack has to happen on this it is easy to know if things will keep working with those changes. I see that many of our test fixtures use net-id maybe some should be updated to use the new setup.

We can also update the devstack plugin config to use an explicit network setup in the nodepool.yaml file.

Patch-set: 1
Label: Code-Review=-1