Update patch set 10
Patch Set 10: Code-Review-1

(2 comments)

We made a deliberate choice to break backwards compatability because we do not want any credentials at all in the nodepool.yaml file.  Doing so is an anti-pattern, and we've already crossed the bridge.  Adding backwards compat is, at this point, merely adding technical debt for no gain.

Patch-set: 10
Label: Code-Review=-1