Update patch set 4
Patch Set 4:

(1 comment)

Proposing an update.  Since the corresponding zuul patch has merged, this is causing an issue with some running their own CI systems (PowerVM for instance is hitting an issue where zuul wants APScheduler > 3.0, but the nodepool on the same node is currently requiring <3.0).

Hoping that version builds on top of what Paul did, but solves the UT issue of patch set 5 and the comment from Clark on patch set 4.

Patch-set: 4
Label: Verified=0