Update patch set 1
Patch Set 1:

I don't know what the solution is, this seems to mostly be an issue for orchestration 

How can we resolve this?  I'm thinking the best argument is just to build consistency.

If I'm correct, status quo with no gear configuration is:

 - builder bails out
 - nodepool says nothing & continues on

with this it would be come

 - builder hangs
 - nodepool says nothing & continues on

I don't think I really care what the behaviour is -- whatever helps orchestration i guess -- but if it was making both do the same thing, I think that's a good argument.

Patch-set: 1