Update patch set 7
Patch Set 7: Code-Review-1

Can you provide some commentary on the notes I left in PS5 that were not addressed?  [And in general, I think if you choose to ignore a potential problem in a review, it's nice to leave an explanatory response.  I don't think you need to respond to everything (actually making a suggested change is sufficient response), and some review comments don't merit a response, but when no action is taken on a serious note it's a bit confusing.]

Patch-set: 7
Label: Code-Review=-1