Handle image upload job failures
We weren't handling image upload job failures which mean that we
orphaned images in the snapshot image table leaving them in a building
state. We should handle failures to upload by removing the building
image from the snapshot image table. Not doing this prevents any new
images with that label from being uploaded to the provider because
nodepool believes there is an upload in progress.

Change-Id: I1874ac187521d3621b37b3ad5cd69f37f04b35a7