Document clouds.yaml as the preferred configuration
Configuring clouds is complicated - possibly more so than is sane, but
hey, that's life. Rather than trying to express all of that directly in
nodepool, the preferred way to deal with it is to define a named cloud
in clouds.yaml and then refer to that in nodepool.

Update the documentation to reflect that.

Change-Id: I1e97dc88227c57cd5f718f8c1fabdcfeafd031c1