Allow jenkins user and apikey to be optional for unauthenticated jenkins as documented
The nodepool docs indicate that authenticated use of Jenkins is optional
and that the user, apikey, and credentials are optional values in
secure.yaml. The code throws an exception when user or apikey are not set.

Expand the try block to include the user and apikey fields along with the credentials_id.

Change-Id: I916eb6adee37cbe3a3dd5985990a63aa30c48805