Add the ability to use ansible for ready scripts
Baking scripts into the base image is a design choice intended to avoid
copying the same code to each host tens of thousands of times per day.
The trade off with that choice is the need to spin new base images any
time a change needs to be landed to ready scripts. As zuul v3 is already
looking towards a future where we'll be using ansible to execute remote
code at run time, it's possible that it's time to allow for the choice
of greater runtime flexibility and responsiveness at the expense of
additional copying.

As the scripts will still be baked in to the image, a hybrid world where
the ready playbook runs and executes some more static and unchanging
bits that are in the scripts dir may wind up being the most
advantageous.

In any case, this adds the ability to have a ready playbook. It depends
on a not-yet-released library which is itself extracted from code in the
zuul v3 branch of zuul.

Change-Id: I5c0f2dfdf6c9d211dd80f1d819aac7ad87064d2a
Signed-off-by: Paul Belanger <pabelanger@redhat.com>