Update patch set 3
Patch Set 3:

Rebased, the configuration handling has been moved from nodepool/nodepool.py to nodepool/config.py (to support secure config).

The only difference is the default value is now '60' when previously it was 1 * MINS , but that MINS constant is only defined in nodepool.py.  I have added a comment to point it is in seconds.

Patch-set: 3